chore: eslint 규칙에 맞게 코드 수정
