+++ src/types/user.ts
@@ -1,0 +1,9 @@
+interface User {
+  id: number
+  name: string
+  username: string
+  isFollowing: boolean
+  image?: string
+}
+
+export type { User }
